Club creation with time Interval working
