--- MovieClub/Views/Clubs/CreateClubForm.swift
@@ -1,25 +1,0 @@
-//
-//  CreateClubForm.swift
-//  MovieClub
-//
-//  Created by Marcus Lair on 5/14/24.
-//
-
-import SwiftUI
-import PhotosUI
-
-struct CreateClubForm: View {
-    @Environment(DataManager.self) var data: DataManager
-    var body: some View {
-        EditEmptyView()
-        
-    }
-}
-    
-    
-
-
-#Preview {
-    CreateClubForm()
-        .environment(DataManager())
-}
